Files of Python scripts to install.
These three files are scripts used to run the Time(s)Splitter. The project is composed of a file called "oldest" and "newest"—these files run the initial query to get, split, and write text to a CSV. The "new material" file allows you to concatenate fragments of strings between the two CSVs. I believe the project may be adapted to incorporate any number of text queries and their concatenations, but this form only considers working with two sets.